backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
requests>=2.32.3
python-dotenv>=1.0.1
openai>=1.57.0
python-dateutil>=2.9.0
